migrates discovery logic
