+++ insecure/corruptlibp2p/spammerGossipSub_test.go
@@ -1,0 +1,45 @@
+package corruptlibp2p
+
+import (
+	"context"
+	"testing"
+	"time"
+
+	"github.com/onflow/flow-go/model/flow"
+	"github.com/onflow/flow-go/module/irrecoverable"
+	"github.com/onflow/flow-go/network/p2p"
+	p2ptest "github.com/onflow/flow-go/network/p2p/test"
+	"github.com/onflow/flow-go/utils/unittest"
+)
+
+func TestSpammerGossipSub(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	signalerCtx := irrecoverable.NewMockSignalerContext(t, ctx)
+	sporkId := unittest.IdentifierFixture()
+	defer cancel()
+
+	count := 5
+	nodes := make([]p2p.LibP2PNode, 0, 5)
+	ids := flow.IdentityList{}
+	inbounds := make([]chan string, 0, 5)
+
+	for i := 0; i < count; i++ {
+		handler, inbound := p2ptest.StreamHandlerFixture(t)
+		node, id := p2ptest.NodeFixture(
+			t,
+			sporkId,
+			t.Name(),
+			p2ptest.WithRole(flow.RoleConsensus),
+			p2ptest.WithDefaultStreamHandler(handler),
+		)
+
+		nodes = append(nodes, node)
+		ids = append(ids, &id)
+		inbounds = append(inbounds, inbound)
+	}
+
+	p2ptest.StartNodes(t, signalerCtx, nodes, 1*time.Second)
+	defer p2ptest.StopNodes(t, nodes, cancel, 1*time.Second)
+
+	p2ptest.LetNodesDiscoverEachOther(t, ctx, nodes, ids)
+}
